+++ Pitch Perfect/Pitch Perfect/Base.lproj/Main.storyboard
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="11762" systemVersion="16E195" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES" initialViewController="BYZ-38-t0r">
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="11762" systemVersion="16E195" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES" initialViewController="BFb-iA-RbF">
     <device id="retina4_7" orientation="portrait">
         <adaptation id="fullscreen"/>
     </device>
@@ -9,10 +9,10 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <scenes>
-        <!--View Controller-->
+        <!--Record Sounds View Controller-->
         <scene sceneID="tne-QT-ifu">
             <objects>
-                <viewController id="BYZ-38-t0r" customClass="ViewController" customModule="Pitch_Perfect" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController id="BYZ-38-t0r" customClass="RecordSoundsViewController" customModule="Pitch_Perfect" customModuleProvider="target" sceneMemberID="viewController">
                     <layoutGuides>
                         <viewControllerLayoutGuide type="top" id="y3c-jy-aDJ"/>
                         <viewControllerLayoutGuide type="bottom" id="wfy-db-euE"/>
@@ -39,6 +39,7 @@
                                 <state key="normal" image="Stop"/>
                                 <connections>
                                     <action selector="stopRecording:" destination="BYZ-38-t0r" eventType="touchUpInside" id="l5i-1n-OAf"/>
+                                    <segue destination="Rjb-1s-HYG" kind="show" id="Vu2-mG-Tmw"/>
                                 </connections>
                             </button>
                         </subviews>
@@ -52,14 +53,50 @@
                             <constraint firstItem="g0q-0A-R4D" firstAttribute="centerX" secondItem="g8b-UJ-dfq" secondAttribute="centerX" id="tol-JN-pGt"/>
                         </constraints>
                     </view>
+                    <navigationItem key="navigationItem" id="wjy-gO-qum"/>
                     <connections>
                         <outlet property="recordButton" destination="hC4-zD-rRQ" id="qIw-3J-DGR"/>
                         <outlet property="recordingLabel" destination="g8b-UJ-dfq" id="Eqz-Bn-xgM"/>
-                        <outlet property="stopRecording" destination="Hfe-c7-VLM" id="OjV-FV-hkm"/>
                         <outlet property="stopRecordingButton" destination="g0q-0A-R4D" id="joe-Rd-akf"/>
                     </connections>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="dkx-z0-nzr" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="726" y="124"/>
+        </scene>
+        <!--View Controller-->
+        <scene sceneID="fOZ-ZX-zz1">
+            <objects>
+                <viewController id="Rjb-1s-HYG" sceneMemberID="viewController">
+                    <layoutGuides>
+                        <viewControllerLayoutGuide type="top" id="tqk-Mf-eIk"/>
+                        <viewControllerLayoutGuide type="bottom" id="Wxt-Ag-XZb"/>
+                    </layoutGuides>
+                    <view key="view" contentMode="scaleToFill" id="faK-QK-tNZ">
+                        <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <color key="backgroundColor" red="0.0" green="1" blue="1" alpha="1" colorSpace="calibratedRGB"/>
+                    </view>
+                </viewController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="BQ7-ZK-HLU" userLabel="First Responder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="1530" y="124"/>
+        </scene>
+        <!--Navigation Controller-->
+        <scene sceneID="fHT-eM-Mil">
+            <objects>
+                <navigationController automaticallyAdjustsScrollViewInsets="NO" id="BFb-iA-RbF" sceneMemberID="viewController">
+                    <toolbarItems/>
+                    <navigationBar key="navigationBar" contentMode="scaleToFill" id="TJH-JF-cff">
+                        <rect key="frame" x="0.0" y="0.0" width="375" height="44"/>
+                        <autoresizingMask key="autoresizingMask"/>
+                    </navigationBar>
+                    <nil name="viewControllers"/>
+                    <connections>
+                        <segue destination="BYZ-38-t0r" kind="relationship" relationship="rootViewController" id="X2q-42-JuW"/>
+                    </connections>
+                </navigationController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="iQA-gY-F2f" userLabel="First Responder" sceneMemberID="firstResponder"/>
             </objects>
             <point key="canvasLocation" x="-98.400000000000006" y="123.68815592203899"/>
         </scene>
