+++ Pitch Perfect/Pitch Perfect/RecordSoundsViewController.swift
@@ -1,0 +1,57 @@
+//
+//  RecordSoundsViewController.swift
+//  Pitch Perfect
+//
+//  Created by Tomas Sidenfaden on 5/16/17.
+//  Copyright © 2017 Tomas Sidenfaden. All rights reserved.
+//
+
+import UIKit
+import AVFoundation
+
+class RecordSoundsViewController: UIViewController {
+
+    var audioRecorder: AVAudioRecorder!
+    
+    @IBOutlet weak var recordingLabel: UILabel!
+    @IBOutlet weak var recordButton: UIButton!
+    @IBOutlet weak var stopRecordingButton: UIButton!
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        stopRecordingButton.isEnabled = false
+    }
+    
+    override func viewWillAppear(_ animated: Bool) {
+        super.viewWillAppear(animated)
+    }
+
+    @IBAction func recordAudio(_ sender: Any) {
+        recordButton.isEnabled = false
+        stopRecordingButton.isEnabled = true
+        recordingLabel.text = "Recording"
+        let dirPath = NSSearchPathForDirectoriesInDomains(.documentDirectory,.userDomainMask, true)[0] as String
+        let recordingName = "recordedVoice.wav"
+        let pathArray = [dirPath, recordingName]
+        let filePath = URL(string: pathArray.joined(separator: "/"))
+        print(filePath)
+        let session = AVAudioSession.sharedInstance()
+        try! session.setCategory(AVAudioSessionCategoryPlayAndRecord, with:AVAudioSessionCategoryOptions.defaultToSpeaker)
+        
+        try! audioRecorder = AVAudioRecorder(url: filePath!, settings: [:])
+        audioRecorder.isMeteringEnabled = true
+        audioRecorder.prepareToRecord()
+        audioRecorder.record()
+    }
+
+    @IBAction func stopRecording(_ sender: Any) {
+        recordButton.isEnabled = true
+        stopRecordingButton.isEnabled = true
+        recordingLabel.text =
+        "Tap to Record"
+        audioRecorder.stop()
+        let audioSession = AVAudioSession.sharedInstance()
+        try! audioSession.setActive(false)
+    }
+}
+
