--- Pitch Perfect/Pitch Perfect/ViewController.swift
@@ -1,39 +1,0 @@
-//
-//  ViewController.swift
-//  Pitch Perfect
-//
-//  Created by Tomas Sidenfaden on 5/16/17.
-//  Copyright © 2017 Tomas Sidenfaden. All rights reserved.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    @IBOutlet weak var recordingLabel: UILabel!
-    @IBOutlet weak var recordButton: UIButton!
-    @IBOutlet weak var stopRecordingButton: UIButton!
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        stopRecordingButton.isEnabled = false
-    }
-    
-    override func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
-    }
-
-    @IBAction func recordAudio(_ sender: Any) {
-        recordButton.isEnabled = false
-        stopRecordingButton.isEnabled = true
-        recordingLabel.text = "Recording"
-    }
-
-    @IBAction func stopRecording(_ sender: Any) {
-        recordButton.isEnabled = true
-        stopRecordingButton.isEnabled = true
-        recordingLabel.text =
-        "Tap to Record"
-    }
-}
-
